refactor xdd fetch function into the data service
